requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Farasa==0.0.1
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
pydub==0.25.1
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
setuptools==75.2.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3